### 2023-04-27 #5414 @kthyng
Hi @miniufo and thanks for your submission! To start, can you take a look at the missing DOIs listed above and if they are accurate, add them into the appropriate references.

Also I see that you have tests, docs, and installation instructions, so that's good. However, your installation through pip is not yet supported — when do you plan to have that set up?
