+++ paper/paper.bib
@@ -8,6 +8,7 @@
   publisher = {Ubiquity Press},
   pages     = {2067-2078},
 	url       = {https://doi.org/10.1002/qj.337},
+  doi       = {10.1002/qj.337},
 }
 
 
@@ -32,6 +33,7 @@
   publisher = {Ubiquity Press},
   pages     = {447-462},
 	url       = {https://doi.org/10.1002/qj.49710644905},
+  doi       = {10.1002/qj.49710644905},
 }
 
 
@@ -45,6 +47,7 @@
   publisher = {Ubiquity Press},
   pages     = {25-43},
 	url       = {https://doi.org/10.2151/jmsj1965.44.1_25},
+  doi       = {10.2151/jmsj1965.44.1_25},
 }
 
 @article{Munk:1950,
@@ -57,6 +60,7 @@
   publisher = {Ubiquity Press},
   pages     = {80-93},
 	url       = {https://doi.org/10.1175/1520-0469(1950)007<0080:OTWDOC>2.0.CO;2},
+  doi       = {10.1175/1520-0469(1950)007<0080:OTWDOC>2.0.CO;2},
 }
 
 
@@ -70,6 +74,7 @@
   publisher = {Ubiquity Press},
   pages     = {202-206},
 	url       = {https://doi.org/10.1029/TR029i002p00202},
+  doi       = {10.1029/TR029i002p00202},
 }
 
 
@@ -83,6 +88,7 @@
   publisher = {Ubiquity Press},
   pages     = {31-38},
 	url       = {https://doi.org/10.1002/qj.49710443903},
+  doi       = {10.1002/qj.49710443903},
 }
 
 
@@ -96,6 +102,7 @@
   publisher = {Ubiquity Press},
   pages     = {877-946},
 	url       = {https://doi.org/10.1002/qj.49711147002},
+  doi       = {10.1002/qj.49711147002},
 }
 
 
@@ -115,8 +122,6 @@
 @article{Rocklin:2015,
 	title   = {Dask: {Parallel} {Computation} with {Blocked} algorithms and {Task} {Scheduling}},
 	shorttitle = {Dask},
-	url     = {http://doi.org/10.25080/Majora-7b98e3ed-013},
-	doi     = {10.25080/majora-7b98e3ed-013},
 	urldate = {2021-11-05},
 	journal = {Proceedings of the 14th Python in Science Conference},
 	author  = {Rocklin, Matthew},
@@ -124,6 +129,8 @@
 	note    = {Conference Name: Proceedings of the 14th Python in Science Conference},
 	pages   = {126--132},
 	file    = {Full Text PDF:/Users/noraloose/Zotero/storage/A3H5QUZ6/Rocklin - 2015 - Dask Parallel Computation with Blocked algorithms.pdf:application/pdf;Snapshot:/Users/noraloose/Zotero/storage/DP4JWDBB/matthew_rocklin.html:text/html},
+	url     = {http://doi.org/10.25080/Majora-7b98e3ed-013},
+	doi     = {10.25080/majora-7b98e3ed-013},
 }
 
 
@@ -134,4 +141,5 @@
   pages     = { 1 - 6 },
   year      = 2015,
 	url       = {https://doi.org/10.1145/2833157.2833162},
+  doi       = {10.1145/2833157.2833162},
 }
