+++ paper/paper.bib
@@ -7,7 +7,7 @@
   year      = 2008,
   publisher = {Ubiquity Press},
   pages     = {2067-2078},
-	url       = {https://doi.org/10.1002/qj.337},
+  url       = {https://doi.org/10.1002/qj.337},
   doi       = {10.1002/qj.337},
 }
 
@@ -32,7 +32,7 @@
   year      = {1980},
   publisher = {Ubiquity Press},
   pages     = {447-462},
-	url       = {https://doi.org/10.1002/qj.49710644905},
+  url       = {https://doi.org/10.1002/qj.49710644905},
   doi       = {10.1002/qj.49710644905},
 }
 
@@ -46,7 +46,7 @@
   year      = 1966,
   publisher = {Ubiquity Press},
   pages     = {25-43},
-	url       = {https://doi.org/10.2151/jmsj1965.44.1_25},
+  url       = {https://doi.org/10.2151/jmsj1965.44.1_25},
   doi       = {10.2151/jmsj1965.44.1_25},
 }
 
@@ -59,7 +59,7 @@
   year      = 1950,
   publisher = {Ubiquity Press},
   pages     = {80-93},
-	url       = {https://doi.org/10.1175/1520-0469(1950)007<0080:OTWDOC>2.0.CO;2},
+  url       = {https://doi.org/10.1175/1520-0469(1950)007<0080:OTWDOC>2.0.CO;2},
   doi       = {10.1175/1520-0469(1950)007<0080:OTWDOC>2.0.CO;2},
 }
 
@@ -73,7 +73,7 @@
   year      = 1948,
   publisher = {Ubiquity Press},
   pages     = {202-206},
-	url       = {https://doi.org/10.1029/TR029i002p00202},
+  url       = {https://doi.org/10.1029/TR029i002p00202},
   doi       = {10.1029/TR029i002p00202},
 }
 
@@ -87,7 +87,7 @@
   year      = 1978,
   publisher = {Ubiquity Press},
   pages     = {31-38},
-	url       = {https://doi.org/10.1002/qj.49710443903},
+  url       = {https://doi.org/10.1002/qj.49710443903},
   doi       = {10.1002/qj.49710443903},
 }
 
@@ -101,7 +101,7 @@
   year      = 1985,
   publisher = {Ubiquity Press},
   pages     = {877-946},
-	url       = {https://doi.org/10.1002/qj.49711147002},
+  url       = {https://doi.org/10.1002/qj.49711147002},
   doi       = {10.1002/qj.49711147002},
 }
 
@@ -115,7 +115,7 @@
   year      = 2017,
   publisher = {Ubiquity Press},
   doi       = {10.5334/jors.148},
-	url       = {https://doi.org/10.5334/jors.148},
+  url       = {https://doi.org/10.5334/jors.148},
 }
 
 
@@ -129,8 +129,8 @@
 	note    = {Conference Name: Proceedings of the 14th Python in Science Conference},
 	pages   = {126--132},
 	file    = {Full Text PDF:/Users/noraloose/Zotero/storage/A3H5QUZ6/Rocklin - 2015 - Dask Parallel Computation with Blocked algorithms.pdf:application/pdf;Snapshot:/Users/noraloose/Zotero/storage/DP4JWDBB/matthew_rocklin.html:text/html},
-	url     = {http://doi.org/10.25080/Majora-7b98e3ed-013},
-	doi     = {10.25080/majora-7b98e3ed-013},
+  url     = {http://doi.org/10.25080/Majora-7b98e3ed-013},
+  doi     = {10.25080/majora-7b98e3ed-013},
 }
 
 
@@ -140,6 +140,6 @@
[…]
