### 2023-05-01 #5414 @kthyng
> The pypi is already available at https://pypi.org/project/xinvert/. So I guess pip installation is available?

Yep looks like it!
