### 2023-05-30 #5414 @ashwinvis
Hi @jbytecode I tried to make room for JOSS, but it seems like I will be occupied most of summer. Otherwise it does seem like an interesting library. Hope you find someone.
