### 2023-05-31 #5414 @jbytecode
👋👋👋 Dear @NoraLoose    👋👋👋

JOSS publishes articles about open source research software. The submission I'd like you to review is titled: "**xinvert: A Python package for inversion problems in geophysical fluid dynamics**". You can find more information at the top of this Github issue (https://github.com/openjournals/joss-reviews/issues/5414).

This is the pre-review issue. After setting at least 2 reviewers we will start the review process in a separate thread. In that thread, there will be ~25 check items for each single reviewer.

We've already assigned a reviewer. If you accept being a reviewer, we can go next.

Thank you in advance!
