### 2023-06-01 #5510 @editorialbot
[REVIEW]: xinvert: A Python package for inversion problems in geophysical fluid dynamics
**Submitting author:** @miniufo (<a href="http://orcid.org/0000-0001-5660-7619">Yu-Kun Qian</a>)
**Repository:** https://github.com/miniufo/xinvert
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.5
