### 2023-06-02 miniufo/xinvert#16 @DamienIrving
JOSS review: Installation
@miniufo - well done on your submission to JOSS.

As part of my review of your submission (see https://github.com/openjournals/joss-reviews/issues/5510), I have some questions/suggestions regarding software installation:
- In the "how to install" (https://github.com/miniufo/xinvert#2-how-to-install) section of your README file it says "install via pip (not yet)" but as far as I can tell it can be installed by pip. What is the meaning of "not yet"?
- Do you think it would be advantageous to also make the package available via conda (i.e. conda-forge) for users who manage their python environments using conda?
