### 2023-06-02 miniufo/xinvert#17 @DamienIrving
JOSS review: Automated tests
@miniufo - well done on your submission to JOSS.

As it stands, your package doesn't meet the following requirement on the reviewer checklist:

> Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?

(See https://github.com/openjournals/joss-reviews/issues/5510 for details.)

You've got some test code available at `tests/` and associated data files at `Data/`, but there is no guidance provided on how to run those tests manually or better still there is no automated testing.

In addition, it looks like your test code culminates in the plotting of figures. While that can be useful for visual verification, common practice would also include a suite of tests that have the following components:
- a *fixture*, which is the thing being tested (e.g., an input data array);
- an *actual result*, which is what the xinvert package produces when given the fixture; and
- an *expected result* that the actual result is compared to.

The comparison between the actual result and expected result would take the form of an assertion, so that the test can be said to pass or fail. You can then use a test runner like pytest in conjunction with [GitHub actions](https://github.com/features/actions) to setup automated testing.

You may already be familiar with these testing concepts, but if not let me know and I can point you to some resources that explain how to setup automated testing.
