### 2023-06-02 miniufo/xinvert#18 @DamienIrving
JOSS review: community guidelines
@miniufo - well done on your submission to JOSS.

As it stands, your package doesn't meet the following requirement on the reviewer checklist:

> Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

(See https://github.com/openjournals/joss-reviews/issues/5510 for details.)
