### 2023-06-02 miniufo/xinvert#19 @DamienIrving
JOSS review: state of the field
@miniufo - well done on your submission to JOSS.

There is a requirement on the reviewer checklist relating to the software paper which reads as follows:
> State of the field: Do the authors describe how this software compares to other commonly-used packages?

(See https://github.com/openjournals/joss-reviews/issues/5510 for details.)

You make a comparison against windspharm in your README file...
> Inversion on the spherical earth, like some meteorological problems, could utilize the spherical harmonics like [windspharm](https://github.com/ajdawson/windspharm), which would be more efficient using FFT than SOR used here. However, in the case of ocean, SOR method is definitely a better choice in the presence of irregular land/sea mask.

... but not in the software paper itself. I think it would be worth mentioning windspharm (and any other relevant commonly-used packages) in your software paper too.
