### 2023-06-02 #5510 @DamienIrving
I've made an initial review of the package.

In general, I think the author has done a great job putting the package together and it has the potential to be a useful resource for the community.

There's just a few items on the checklist that I don't think can be checked off at the moment. I have opened issues at the xinvert repository relating to each of those items.
