### 2023-06-02 miniufo/xinvert#17 @miniufo
You are right, I know this concept, probably using `assert` keyword in the program, but not every detail.  Could you please send me some concrete examples that I can follow?

Since it is a numerical solver, I cannot expect the repeated runs of the codes reproduce the same results.  I only visualize the plots to see if they are close to those plots in the related papers.
