### 2023-06-02 miniufo/xinvert#16 @miniufo
- that's right.  I forgot to update this.  (not yet) should be removed.
- Definitely should be in conda.  Right now I send a PR to conda for another package [here](https://github.com/conda-forge/staged-recipes/pull/22849), which is the first time I do this.  I am not sure how long it would take (or maybe I do something incorrect).  Maybe you can point me to the right path on how to make `xinvert` available via conda.
