### 2023-06-05 miniufo/xinvert#17 @DamienIrving
This chapter in *Research Software Engineering with Python* gives an overview of general testing concepts:
https://merely-useful.tech/py-rse/testing.html

This page gives more information on using pytest:
https://realpython.com/pytest-python-testing/

In terms of a concrete example, here's a test file I wrote recently:
https://github.com/climate-innovation-hub/qqscale/blob/master/test_qdm.py

While you can't expect repeated runs to reproduce exactly the same results, you should hopefully be able to create a fixture that produces an expected result within some tolerance (i.e. `assert np.allclose(expected_result, actual_result)`).

Once you've got your tests setup, you can use GitHub Actions to run the tests every time you make a commit to your repository. That involves creating a `tests.yml` file in a new `.github/workflows/` directory. It will look something like:
https://github.com/climate-innovation-hub/qqscale/blob/master/.github/workflows/tests.yml

Hope that helps :smile:
