### 2023-06-05 miniufo/xinvert#16 @DamienIrving
Your xgrads PR looks fine to me - hopefully they will action it soon.

Since xinvert is available via pip I don't think waiting for it to be available via conda should delay your JOSS publication. It would be nice to have it available via conda eventually, but it's not absolutely necessary for publication with JOSS.
