### 2023-06-05 miniufo/xinvert#16 @miniufo
That's great.  I'll follow the steps of xgrads and make a PR of adding xinvert to conda recipe.
