### 2023-06-05 miniufo/xinvert#17 @miniufo
Thanks very much for your kind help.  I will follow these to make auto-tests for xinvert soon.
