+++ docs/source/Contributors.rst
@@ -1,0 +1,23 @@
+.. xinvert documentation master file, created by
+   sphinx-quickstart on Wed April 19 21:26:54 2023.
+   You can adapt this file completely to your liking, but it should at least
+   contain the root `toctree` directive.
+
+Contributors
+============
+
+`xinvert` is meant to be a open-source package and we welcome feedback and contributions.
+
+You can [open an issue in the github issues page](https://github.com/miniufo/xinvert/issues) if you:
+
+- find a bug in the package;
+- need some new features;
+- want some supports;
+- have your own examples.
+
+If you have a cool or new example that can be fit into `xinvert`, please consider contributing to [Examples](xinvert/docs/source/notebooks) in a form of Jupyter Notebook.  Generally, the notebook contains:
+1. an introduction of the problem;
+2. a mathematical formulation of the problem into a elliptic equations;
+3. results from inversion and associated demonstration.
+
+In order to contribute to `xinvert`, please fork the repository, modify the codes, and send a pull request (PR).  We are looking forward to your contributions.
