+++ docs/source/Examples.rst
@@ -21,5 +21,7 @@ problems can be found at `this notebook <https://github.com/miniufo/xinvert/blob
    notebooks/04_Eliassen_model
    notebooks/07_Gill_Matsuno_model
    notebooks/08_Stommel_Munk_model
-   notebooks/09_Omega_equation
-   notebooks/10_3D_Ocean_flow
+   notebooks/09_Fofonoff_flow
+   notebooks/10_2D_Steady_flow_over_topography
+   notebooks/11_Omega_equation
+  
