+++ tests/testFDs.py
@@ -1,159 +1,0 @@
-# -*- coding: utf-8 -*-
-"""
-Created on 2020.12.25
-
-@author: MiniUFO
-Copyright 2018. All rights reserved. Use is subject to license terms.
-"""
-#%% load sample data
-import xarray as xr
-
-
-dset = xr.open_dataset('./xinvert/Data/Helmholtz_atmos.nc')
-
-T = dset.sf[0]
-
-print(T)
-
-#%% test padding
-from xinvert.xinvert.finitediffs import padBCs
-
-T_Px = padBCs(T, dim='lon', BCs=('fixed','fixed'), fill=(1,1))
-T_Py = padBCs(T, dim='lat', BCs=('extend','fixed'), fill=(2,2))
-T_Py2= padBCs(T, dim='lat', BCs=('periodic','periodic'))
-T_Py3= padBCs(T, dim='lat', BCs=('reflect','extend'), fill=(3,3))
-
-
-#%% test derivatives
-from xinvert.xinvert.finitediffs import deriv, deriv2
-
-Tx1 = deriv(T, dim='lon', scheme='center')
-Tx2 = deriv(T, dim='lon', scheme='forward')
-Tx3 = deriv(T, dim='lon', scheme='backward')
-
-Txx = deriv2(T, dim='lon')
-
-#%% test grad, curl, vort, Laplacian
-from xinvert.xinvert import FiniteDiff
-
-fd = FiniteDiff(dim_mapping={'T':'time', 'Y':'lat', 'X':'lon'},
-                BCs={'Y':'reflect', 'X':'periodic'},
-                coords='lat-lon')
-
-Ty, Tx = fd.grad(T, dims=['Y', 'X'])
-Tcurl  = fd.curl(Tx, Ty)
-Tdivg  = fd.divg([Tx, Ty], dims=['X', 'Y'])
-TLap   = fd.Laplacian(T, dims=['Y', 'X'])
-
-#%%
-from GeoApps.GridUtils import add_latlon_metrics
-from GeoApps.DiagnosticMethods import Dynamics
-
-
-ds, grid = add_latlon_metrics(dset, {'Y':'lat', 'X':'lon'},
-                              boundary={'X':'periodic', 'Y':'extend'})
-
-dyn = Dynamics(ds, grid, arakawa='A')
-
-Ty2, Tx2= dyn.grad(T, dims=['Y', 'X'])
-Tcurl2  = dyn.vort(u=Tx2, v=Ty2, components='k')
-Tdivg2  = dyn.divg([Tx2, Ty2], dims=['X', 'Y'])
-TLap2   = dyn.Laplacian(T, dims=['X', 'Y'])
-
-
-#%%
-import proplot as pplt
-import numpy as np
-
-fig, axes = pplt.subplots(nrows=5, ncols=3, figsize=(11, 12))
-
-fontsize = 14
-
-ax = axes[0,0]
-m=ax.contourf(Tx, levels=np.linspace(-35, 35, 15))
-ax.set_title('Tx by FiniteDiff', fontsize=fontsize)
-ax = axes[0,1]
-m=ax.contourf(Tx2, levels=np.linspace(-35, 35, 15))
-ax.colorbar(m, loc='r')
-ax.set_title('Tx2 by xgcm', fontsize=fontsize)
-ax = axes[0,2]
-m=ax.contourf(Tx-Tx2, levels=21)
-ax.colorbar(m, loc='r')
-ax.set_title('diff of Tx', fontsize=fontsize)
-
-ax = axes[1,0]
-m=ax.contourf(Ty, levels=np.linspace(-35, 35, 15))
[…]
