+++ docs/source/Contributors.rst
@@ -8,14 +8,14 @@ Contributors
 
 `xinvert` is meant to be a open-source package and we welcome feedback and contributions.
 
-You can [open an issue in the github issues page](https://github.com/miniufo/xinvert/issues) if you:
+You can `open an issue in the github issues page <https://github.com/miniufo/xinvert/issues>`__ if you:
 
 - find a bug in the package;
 - need some new features;
 - want some supports;
 - have your own examples.
 
-If you have a cool or new example that can be fit into `xinvert`, please consider contributing to [Examples](xinvert/docs/source/notebooks) in a form of Jupyter Notebook.  Generally, the notebook contains:
+If you have a cool or new example that can be fit into `xinvert`, please consider contributing to `Examples <./xinvert/docs/source/notebooks>`__ in a form of Jupyter Notebook.  Generally, the notebook contains:
 1. an introduction of the problem;
 2. a mathematical formulation of the problem into a elliptic equations;
 3. results from inversion and associated demonstration.
