+++ docs/source/Contributors.rst
@@ -15,9 +15,16 @@ You can `open an issue in the github issues page <https://github.com/miniufo/xin
 - want some supports;
 - have your own examples.
 
-If you have a cool or new example that can be fit into `xinvert`, please consider contributing to `Examples <./xinvert/docs/source/notebooks>`__ in a form of Jupyter Notebook.  Generally, the notebook contains:
-1. an introduction of the problem;
-2. a mathematical formulation of the problem into a elliptic equations;
-3. results from inversion and associated demonstration.
+If you have a cool or new example that can be fit into `xinvert`, please consider contributing to `Examples <https://github.com/miniufo/xinvert/tree/master/docs/source/notebooks>`__ in a form of Jupyter Notebook.  Generally, the notebook contains:
 
-In order to contribute to `xinvert`, please fork the repository, modify the codes, and send a pull request (PR).  We are looking forward to your contributions.
+- an introduction of the problem;
+- a mathematical formulation of the problem into a elliptic equations;
+- results from inversion and associated demonstration.
+
+In order to contribute to `xinvert`, please
+
+- fork the repository,
+- modify the codes, and
+- send a pull request (PR).
+
+We are looking forward to your contributions.
