+++ requirements.txt
@@ -2,4 +2,5 @@ xarray
 dask
 scipy
 numba
+netcdf4
 
