+++ requirements.txt
@@ -2,5 +2,5 @@ xarray
 dask
 scipy
 numba
-netcdf4
+netCDF4
 
