+++ .github/workflows/tests.yml
@@ -23,6 +23,4 @@ jobs:
 
       - name: Test with pytest
         run: |
-          pwd
-          ls -l ./xinvert/Data/
           python -m pytest
