+++ tests/test_AnimateConverge.py
@@ -11,7 +11,7 @@ from xinvert import animate_iteration
 
 
 def test_animate():
-    ds = xr.open_dataset('../Data/Helmholtz_atmos.nc')
+    ds = xr.open_dataset('./Data/Helmholtz_atmos.nc')
     
     assert ds.dims == {'time': 2, 'lat': 73, 'lon': 144}
     
