+++ tests/test_Geopotential.py
@@ -11,6 +11,8 @@ import numpy as np
 from xinvert import FiniteDiff, invert_Poisson
 
 def test_Geopotential():
+    import os
+    os.system('pwd')
     ds = xr.open_dataset('./xinvert/Data/atmos3D.nc').sel({'LEV':500})
     
     u = ds.U
