+++ tests/test_Geopotential.py
@@ -11,8 +11,14 @@ import numpy as np
 from xinvert import FiniteDiff, invert_Poisson
 
 def test_Geopotential():
-    import os
-    os.system('pwd')
+    from os.path import exists
+    print('testing relative paths')
+    print(exists('./Data/atmos3D.nc'))
+    print(exists('../Data/atmos3D.nc'))
+    print(exists('../../Data/atmos3D.nc'))
+    print(exists('./xinvert/Data/atmos3D.nc'))
+    print(exists('../xinvert/Data/atmos3D.nc'))
+    print(exists('../../xinvert/Data/atmos3D.nc'))
     ds = xr.open_dataset('./xinvert/Data/atmos3D.nc').sel({'LEV':500})
     
     u = ds.U
