### 2023-06-09 miniufo/xinvert#17 @miniufo
Hi @DamienIrving, I just followed your guide and made a action for tests.  Everything works fine locally on my laptop.  But the action build raised an error: **did not find a match in any of xarray's currently installed IO backends ['netcdf4', 'scipy']**

I made both `netcdf4` and `scipy` in the `requirement.txt` file, so I don't know how to resolve this.  Could you help me find why this is happening?

PS: I use `git lfs` this time because some of the sample netcdf files are larger than github limits.
