### 2023-06-11 miniufo/xinvert#17 @DamienIrving
If you add an extra step in `tests.yml` after the "install dependencies" step to list the installed packages you'll see that netCDF4 is successfully installed, so that isn't the problem.
```
      - name: List installed packages
        shell: bash -l {0}
        run: pip list
```

It might be that the files are too large... you could try a test with a very small data file?
