### 2023-06-11 miniufo/xinvert#17 @miniufo
Hi @DamienIrving, thanks for your tries and suggestion.  I guess I found the reason.  All netcdf files in `Data` has been tracked by LFS, which are then not recognized by `xr.open_dataset`.  I modified a small nc file but leave it outside of LFS, then it is passed (you may check the log).

So this is really embarrassing, because I need some global dataset to make a test, which exceeds the limit by github.  LFS seems to upload big file transparently but not openable online.  Do you have any suggestion?
