### 2023-06-12 miniufo/xinvert#17 @DamienIrving
It sounds like you need to refactor the tests you want GitHub Actions to run so they don't require large input datasets.

For automated testing, it's most common to either have the test code create it's own fake data (e.g. https://github.com/climate-innovation-hub/qqscale/blob/master/test_qdm.py#L15) so that you don't need to include any data files in your repo, or if using "real" data is unavoidable then it's common to use a netCDF file containing only a single grid point or small spatial area.

You can do your own testing offline where you process and plot large global datasets, but those offline tests wouldn't be included in the tests run by GitHub Actions.

Does that make sense?
