### 2023-06-12 miniufo/xinvert#17 @miniufo
Well, that's right.  I should skip those tests requiring large datasets (>100MB), and do them locally by hands.
