### 2023-06-12 miniufo/xinvert#19 @miniufo
I've add one paragraph [here](https://github.com/miniufo/xinvert/blob/master/paper/paper.md#state-of-the-field) to address the state of the field.  Please see if it is OK now.
