### 2023-06-12 miniufo/xinvert#18 @miniufo
I've added a page to the doc [here](https://xinvert.readthedocs.io/en/latest/Contributors.html).  Please see if this is OK.
