+++ .github/workflows/tests.yml
@@ -13,22 +13,24 @@ jobs:
 
     steps:
       - name: Checkout source
-        uses: actions/checkout@v3
+        uses: actions/checkout@v2
 
-      - name: Install dependencies
-        run: |
-          python -m pip install --upgrade pip
-          pip install ruff pytest
-          if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
+      - name: setup conda environment
+        uses: conda-incubator/setup-miniconda@v2
+        with:
+          miniconda-version: "latest"
+          python-version: ${{ matrix.python-version }}
+          environment-file: environment.yml
+          activate-environment: xgrads
+          auto-activate-base: false
 
-      - name: Check something
-        run: |
-          echo ${{ github.workspace }}
+      - name: Check workspace
+        run: echo ${{ github.workspace }}
 
       - name: List installed packages
         shell: bash -l {0}
-        run: pip list
+        run: conda list
 
       - name: Test with pytest
-        run: |
-          python -m pytest
+        shell: bash -l {0}
+        run: python -m pytest
