+++ environment.yml
@@ -1,0 +1,11 @@
+name: xinvert
+channels:
+  - conda-forge
+dependencies:
+  - python
+  - pytest
+  - numpy
+  - xarray
+  - dask
+  - scipy
+  - netCDF4
