### 2023-06-14 miniufo/xinvert#16 @miniufo
Hi @DamienIrving, I've removed **'not yet'** in the [README.md](https://github.com/miniufo/xinvert).  Also, I've sent a PR to [conda recipe](https://github.com/conda-forge/staged-recipes/pull/23082), waiting it to be merged.  Please see if this issue is fixed now.
