### 2023-06-15 miniufo/xinvert#17 @DamienIrving
Thanks for implementing automated testing.

Just so I understand, what are the tests in [test_Poisson.py](https://github.com/miniufo/xinvert/blob/master/tests/test_Poisson.py) checking?

I would expect that each test would have three basic steps:
1. Read in vorticity and divergence data from a file (that's the test "fixture")
2. Calculate the streamfunction and velocity potential (that's the "actual result")
3. Compare the actual result to an "expected result" to see if the test passes (i.e. using an assertion)

Do you have an expected result? If not, is there some vorticity and divergence data that you can create (i.e. simple example data) where you know what the streamfunction and velocity potential values should be?
