### 2023-06-15 miniufo/xinvert#17 @miniufo
Hi, @DamienIrving thanks for your careful check on this.  You are right that one could find some analytical results to do the tests.  But here I adopted an indirect way here.  Basically, we are not checking the streamfunction itself, but its gradient (or velocity components).

According to [Helmholtz decomposition theory](https://en.wikipedia.org/wiki/Helmholtz_decomposition), a vector field, say, $\mathbf u=(u, v)$, can be decomposed into rotational and divergent components $\mathbf u = \mathbf u_r + \mathbf u_d$.  The rotational component fully recovers the vorticity of the original $\mathbf u$ and has exactly zero-divergence, whereas the divergent component fully recovers the divergence of $\mathbf u$ and has exactly zero-vorticity.  So we can theck this to see if the decomposition (i.e., Poisson inverion) is right.

That is why I put `assert`s to see if the divergence of $\mathbf u_r$ is zero, and if the vorticity of $\mathbf u_d$ is zero, **which are expected results**.  The slicing from 1 to -1 is because the finite difference at the boundaries (north-south poles) could result larger errors above the tolerance so it is better to check the interior.  But zonal is periodic so not a problem here.

One of the potential problem with analytical expression is that, we use finite difference scheme here.  If we use larger grid size, there would be grid-scale truncation error relative to analytical results, which may be above the tolerance.  But the current asserts do not depend on grid size, because the inversion for streamfunction, as well we the calculation of its gradient or velocity components, all use the consistent finite difference scheme, which turns out to be quite good that the expected $\nabla\cdot\mathbf u_r=0$ is achieved with default numpy tolerance.

Also, there will be an arbitary constant to Poisson solution.  So any streamfunction, after adding a constant, is still a solution to Poisson equation.  I guess this could be a problem with `windspharm` to determine the arbitary constant.  That is why I do really prefer the **indirect** way of assertion here, as it avoid the arbitary constant.
