### 2023-06-15 miniufo/xinvert#17 @DamienIrving
Thanks for the explanation - that's a nice way to write the tests!
