### 2023-06-15 #5510 @DamienIrving
Update: @miniufo has satisfactorily worked through each of the four issues I opened at the xinvert repo. I think xinvert now satisfies all of the criteria in the review checklist.
