### 2023-06-16 #5510 @jbytecode
@openjournals/dev - pdf generation problem. the error message:

> /home/runner/work/_actions/xuanxu/paper-action/main/vendor/bundle/ruby/3.2.0/gems/theoj-1.5.4/lib/theoj/paper.rb:158:in `failure': Downloading of the repository failed. Please make sure the URL is correct. (Theoj::Error)

what to do for that?
