### 2023-06-16 f1b3b46
remove large nc file
