+++ conda/build.sh
@@ -1,0 +1,2 @@
+echo "Building xinvert to conda"
+$PYTHON setup.py install     # Python command to install the script.
