+++ conda/meta.yaml
@@ -1,0 +1,41 @@
+{% set data = load_setup_py_data() %}
+
+package:
+  name: "xinvert"
+  version: {{ data.get('version') }}
+
+source:
+  git_url: https://github.com/miniufo/xinvert
+
+build:
+  number: 0
+  noarch: python
+
+requirements:
+  host:
+    - python
+    - xarray
+    - dask
+    - numpy
+    - numba
+    - netCDF4
+    - scipy
+  build:
+    - python
+  run:
+    - python
+    - xarray
+    - dask
+    - numpy
+    - numba
+    - netCDF4
+    - scipy
+about:
+  home: https://github.com/miniufo/xinvert
+  license: MIT
+  license_file: LICENSE
+  summary: Invert geophysical fluid dynamic problems (elliptic partial differential equations) using SOR iteration method.
+
+extra:
+  recipe-maintainers:
+    - miniufo
