+++ .travis.yml
@@ -2,7 +2,7 @@ language: python
 script: true
 
 python:
-- "3.6"
+- "3.8"
 
 install:
 - pip install -r requirements.txt
