### 2023-06-30 miniufo/xinvert#21 @NoraLoose
Add installation instructions to the documentation
It would be helpful to have the installation instructions not only in the README, but also in the documentation.

This is part of the JOSS review process at https://github.com/openjournals/joss-reviews/issues/5510.
