### 2023-06-30 miniufo/xinvert#21 @NoraLoose
Also, it would be great if you could remind the user (in installation instructions in README and doc) how to create the conda environment necessary to run the notebooks in the documentation.
```
cd xinvert
conda env create -f environment.yml
```
