### 2023-06-30 miniufo/xinvert#21 @NoraLoose
You could also consider adding `ipykernel` to the list of packages in `environment.yml` so people can use the conda environment on the jupyter hub to run the notebooks. That's up to you!
