### 2023-06-30 miniufo/xinvert#21 @NoraLoose
Further necessary packages to run the jupyter notebooks that could be added to `environment.yml` are:

- [ ] proplot
- [ ] cartopy
- [ ] ?
