### 2023-06-30 miniufo/xinvert#22 @NoraLoose
Software paper: Citation of python packages
To cite the packages windspharm and Dedalus, please use the citations to the software papers (rather than just linking to the package github repos):

- [ ] https://openresearchsoftware.metajnl.com/articles/10.5334/jors.129
- [ ] https://journals.aps.org/prresearch/abstract/10.1103/PhysRevResearch.2.023068

This is part of the JOSS review process at https://github.com/openjournals/joss-reviews/issues/5510.
