### 2023-07-03 miniufo/xinvert#21 @miniufo
Thanks @NoraLoose for these helpful suggestions.

> It would be helpful to have the installation instructions not only in the README, but also in the documentation.

Forgot about this.  I should add the instructions these days to the documentation.

> Also, it would be great if you could remind the user (in installation instructions in README and doc) how to create the conda environment necessary to run the notebooks in the documentation.

I guess I am not familiar with the jupyer-hub things.  So I am not quite sure how to run the notebooks by other people.  Could you please send me a tutorial on the jupyter-hub concept or practice, so that I can figure out a full list of pkgs needed in `environment.yml`?
