### 2023-07-03 miniufo/xinvert#21 @NoraLoose
Sorry, you don't really need the jupyter hub to run the notebooks. If you don't use the jupyter hub, I assume you follow the following workflow:

1) Activate some conda environment.
2) Start jupyter.
3) Run the notebooks in the documentation.

The question is what packages are in the conda environment from 1)? There must be more packages than in [environment.yml](https://github.com/miniufo/xinvert/blob/master/environment.yml), because you need packages like proplot, cartopy, etc.

You could either add these packages to [environment.yml](https://github.com/miniufo/xinvert/blob/master/environment.yml), or make a second environment file, e.g., `environment-notebooks.yml`, which includes all packages from [environment.yml](https://github.com/miniufo/xinvert/blob/master/environment.yml) plus the plotting packages that are only needed for the notebooks. Does this make sense?
