### 2023-07-04 miniufo/xinvert#21 @miniufo
Thanks @NoraLoose.  Don't worry, I really want to learn some new things.  I could add those pkgs I remember.  But sometimes I may miss one or two (actually I didn't remember why there is an `environment.yml` in the folder).  So the best practice is to run it in a new environment (like jupyter hub).  Then I know if I miss any pkg or not.  Maybe all the imported pkgs in the notebooks should be OK?
