### 2023-07-08 miniufo/xinvert#21 @miniufo
Hi @NoraLoose, I really want to add a installation from conda.  I've sent a PR [here](https://github.com/conda-forge/staged-recipes/pull/23082) weeks ago, but no response at all.  Just want to know if I missed something, as this is the first time I do this.  Do I need to ask someone to review the PR?
