### 2023-07-10 miniufo/xinvert#21 @NoraLoose
Hi @miniufo! There are two, distinct conda-related issues:

1. Being able to install xinvert from conda. That is the PR you opened at https://github.com/conda-forge/staged-recipes/pull/23082
2. Providing an environment.yaml file so that users of xinvert can create a conda environment that has all the packages that are necessary to run the notebooks.

I was referring to the second point.

See for example https://gcm-filters.readthedocs.io/en/latest/installation.html#installation-from-conda-forge versus https://gcm-filters.readthedocs.io/en/latest/installation.html#how-to-run-the-example-notebooks as an example for the differences between 1. and 2.
