+++ paper/paper.bib
@@ -120,15 +120,15 @@
 
 
 @article{Rocklin:2015,
-	title   = {Dask: {Parallel} {Computation} with {Blocked} algorithms and {Task} {Scheduling}},
-	shorttitle = {Dask},
-	urldate = {2021-11-05},
-	journal = {Proceedings of the 14th Python in Science Conference},
-	author  = {Rocklin, Matthew},
-	year    = 2015,
-	note    = {Conference Name: Proceedings of the 14th Python in Science Conference},
-	pages   = {126--132},
-	file    = {Full Text PDF:/Users/noraloose/Zotero/storage/A3H5QUZ6/Rocklin - 2015 - Dask Parallel Computation with Blocked algorithms.pdf:application/pdf;Snapshot:/Users/noraloose/Zotero/storage/DP4JWDBB/matthew_rocklin.html:text/html},
+  title   = {Dask: {Parallel} {Computation} with {Blocked} algorithms and {Task} {Scheduling}},
+  shorttitle = {Dask},
+  urldate = {2021-11-05},
+  journal = {Proceedings of the 14th Python in Science Conference},
+  author  = {Rocklin, Matthew},
+  year    = 2015,
+  note    = {Conference Name: Proceedings of the 14th Python in Science Conference},
+  pages   = {126-132},
+  file    = {Full Text PDF:/Users/noraloose/Zotero/storage/A3H5QUZ6/Rocklin - 2015 - Dask Parallel Computation with Blocked algorithms.pdf:application/pdf;Snapshot:/Users/noraloose/Zotero/storage/DP4JWDBB/matthew_rocklin.html:text/html},
   url     = {http://doi.org/10.25080/Majora-7b98e3ed-013},
   doi     = {10.25080/majora-7b98e3ed-013},
 }
@@ -142,4 +142,31 @@
   year      = 2015,
   url       = {https://doi.org/10.1145/2833157.2833162},
   doi       = {10.1145/2833157.2833162},
+}
+
+
+@article{Dawson:2016,
+  title   = {Windspharm: A High-Level Library for Global Wind Field Computations Using Spherical Harmonics},
+  author  = {Dawson, Andrew},
+  journal = {Journal of Open Research Software},
+  volume  = {4},
+  number  = {1},
+  year    = 2016,
+  pages   = {1-5},
+  url     = {https://openresearchsoftware.metajnl.com/articles/10.5334/jors.129},
+  doi     = {10.5334/jors.129},
+}
+
+
+@article{Burns:2020,
+  title     = {Dedalus: A flexible framework for numerical simulations with spectral methods},
+  author    = {Burns, K. J. and Vasil G. M. and Oishi J. S. and Lecoanet, D. and Brown, B. P.},
+  journal   = {Physical Review Research},
+  volume    = {2},
+  number    = {023068},
+  year      = 2020,
+  publisher = {Ubiquity Press},
+  pages     = {1-39},
+  url       = {https://journals.aps.org/prresearch/abstract/10.1103/PhysRevResearch.2.023068},
+  doi       = {10.1103/PhysRevResearch.2.023068},
 }
