+++ docs/source/index.rst
@@ -23,6 +23,7 @@ using its `Zenodo DOI <https://doi.org/10.5281/zenodo.7801501>`__.
    :maxdepth: 2
    :caption: Contents:
    
+   Installation
    notebooks/00_Introduction
    Examples
    Contributors
