+++ paper/paper.bib
@@ -106,6 +106,31 @@
 }
 
 
+@article{Fofonoff:1954,
+  title     = {Steady flow in a frictionless homogeneous ocean},
+  author    = {Fofonoff, N. P.},
+  journal   = {Journal of Marine Research},
+  volume    = {13},
+  year      = 1954,
+  publisher = {Ubiquity Press},
+  pages     = {254-262},
+  url       = {https://elischolar.library.yale.edu/journal_of_marine_research/818/},
+}
+
+
+@article{Bretherton:1976,
+  title     = {Two-dimensional turbulence above topography},
+  author    = {Bretherton, F. P. and Haidvogel, D. B.},
+  journal   = {Journal of Fluid Mechanics},
+  volume    = {78},
+  year      = 1976,
+  publisher = {Ubiquity Press},
+  pages     = {129-154},
+  url       = {https://www.cambridge.org/core/journals/journal-of-fluid-mechanics/article/twodimensional-turbulence-above-topography/839ED65E9B6684B5D6CFEFBD8108A82E},
+  doi       = {10.1017/S002211207600236X},
+}
+
+
 @article{Hoyer:2017,
   title     = {xarray: ND labeled arrays and datasets in Python},
   author    = {Hoyer, Stephan and Hamman, Joe},
@@ -134,17 +159,6 @@
 }
 
 
-@InProceedings{ Lam:2015,
-  author    = { Lam, Siu Kwan and Pitrou, Antoine and Seibert, Stanley },
-  title     = { Numba: a LLVM-based Python JIT compiler },
-  booktitle = { Proceedings of the Second Workshop on the LLVM Compiler Infrastructure in HPC },
-  pages     = { 1 - 6 },
-  year      = 2015,
-  url       = {https://doi.org/10.1145/2833157.2833162},
-  doi       = {10.1145/2833157.2833162},
-}
-
-
 @article{Dawson:2016,
   title   = {Windspharm: A High-Level Library for Global Wind Field Computations Using Spherical Harmonics},
   author  = {Dawson, Andrew},
@@ -169,4 +183,15 @@
   pages     = {1-39},
   url       = {https://journals.aps.org/prresearch/abstract/10.1103/PhysRevResearch.2.023068},
   doi       = {10.1103/PhysRevResearch.2.023068},
+}
+
+
+@InProceedings{ Lam:2015,
+  author    = { Lam, Siu Kwan and Pitrou, Antoine and Seibert, Stanley },
+  title     = { Numba: a LLVM-based Python JIT compiler },
+  booktitle = { Proceedings of the Second Workshop on the LLVM Compiler Infrastructure in HPC },
+  pages     = { 1 - 6 },
+  year      = 2015,
+  url       = {https://doi.org/10.1145/2833157.2833162},
+  doi       = {10.1145/2833157.2833162},
 }
