### 2023-07-11 miniufo/xinvert#21 @miniufo
Sorry @NoraLoose, I know these two issues are different.  The second one is not too hard to fix.  I just prefer to having both issues fixed at the same time, as they are all related to installation here.
