### 2023-07-11 bdec53e
modify environment.yml to run notebook
