### 2023-07-11 miniufo/xinvert#21 @miniufo
Hi @NoraLoose, I've added the installation page to the docs at [here](https://xinvert.readthedocs.io/en/latest/Installation.html) (for conda, I add *not available yet*).  The `environment.yml` is also updated and tested by creating a new env so that the notebooks can be executed.  Please check these again.  Thank you very much.
