### 2023-07-11 miniufo/xinvert#22 @miniufo
Hi @NoraLoose, the citations are updated.  The newly-generated PDF is at [here](https://github.com/openjournals/joss-papers/blob/joss.05510/joss.05510/10.21105.joss.05510.pdf).  Please check this.
