### 2023-07-12 miniufo/xinvert#21 @NoraLoose
> I've added the installation page to the docs at [here](https://xinvert.readthedocs.io/en/latest/Installation.html) (for conda, I add not available yet)

I would delete the [section about installation from conda](https://xinvert.readthedocs.io/en/latest/Installation.html#installation-from-conda). If a user does not read carefully, they may think conda installation is available. You can re-add this to the documentation once installation from conda is working.

Instead, I would add a section on how to create the conda environment to run the example notebooks, similarly to what is done [here](https://gcm-filters.readthedocs.io/en/latest/installation.html#how-to-run-the-example-notebooks).

I just tried your new `environment.yml` and execute all notebooks. I get the error
```
ModuleNotFoundError: No module named 'xcontour'
```
