### 2023-07-13 miniufo/xinvert#21 @miniufo
Hi, @NoraLoose I've removed the part of conda install from both [README](https://github.com/miniufo/xinvert) and [docs page](https://xinvert.readthedocs.io/en/latest/Installation.html).

The dependence of `xcontour` is a bit complicated.  It is another package I've written [here](https://github.com/miniufo/xcontour).  So the example of [reference state](https://github.com/miniufo/xinvert/blob/master/docs/source/notebooks/05_reference_SWM.ipynb), which is newly added, depends on `xcontour` to prepare the forcing (actually preparing the forcing data is not the core part of `xinvert`).  This is not necessary for other examples.  So far, I don't have any plan to get `xcontour` released as a conda package, so it is not possible to add this to conda dependency.  Do you have any suggestion?  I really like this example to remain in the problem list.

By the way, do you know how long my [conda-recipe PR](https://github.com/conda-forge/staged-recipes/pull/23082) will be reviewed and merged?  Since I am doing this first time, I've no idea about this.  It's been a month now.
