### 2023-07-13 miniufo/xinvert#23 @NoraLoose
Dependence on xcontour package
I'm opening a new issue here to continue the discussion on the issue of the xcontour dependence that we started [here](https://github.com/miniufo/xinvert/issues/21#issuecomment-1633983103).

> The dependence of xcontour is a bit complicated. It is another package I've written [here](https://github.com/miniufo/xcontour). So the example of [reference state](https://github.com/miniufo/xinvert/blob/master/docs/source/notebooks/05_reference_SWM.ipynb), which is newly added, depends on xcontour to prepare the forcing (actually preparing the forcing data is not the core part of xinvert). This is not necessary for other examples. So far, I don't have any plan to get xcontour released as a conda package, so it is not possible to add this to conda dependency. Do you have any suggestion? I really like this example to remain in the problem list.

Is there a way to precompute M(Q) and C(Q) with `xcontour` and in the https://xinvert.readthedocs.io/en/latest/notebooks/05_reference_SWM.html example, and just load these states from netcdf if `xcontour` is not available? You could  then adapt the notebook as follows:
```
try:
    import xcontour
    # do xcontour computation in this notebook, as you do now
except ImportError:
    ds = xr.open_dataset(your_precomputed_dataset)
```

This is part of the review process at https://github.com/openjournals/joss-reviews/issues/5510.
