### 2023-07-13 miniufo/xinvert#23 @miniufo
Oh yes, that's a nice workaround.  I should do this follow your suggestion.
