### 2023-07-13 miniufo/xinvert#21 @NoraLoose
> By the way, do you know how long my https://github.com/conda-forge/staged-recipes/pull/23082 will be reviewed and merged? Since I am doing this first time, I've no idea about this. It's been a month now.

Sorry, I'm not sure either.

> Instead, I would add a section on how to create the conda environment to run the example notebooks, similarly to what is done [here](https://gcm-filters.readthedocs.io/en/latest/installation.html#how-to-run-the-example-notebooks).

Are you still planning to do this?
