+++ docs/environment.yml
@@ -1,0 +1,14 @@
+name: xinvert
+channels:
+  - conda-forge
+dependencies:
+  - python
+  - pytest
+  - numpy
+  - xarray
+  - dask
+  - scipy
+  - netCDF4
+  - numba
+  - proplot
+  - cartopy
