### 2023-07-14 miniufo/xinvert#21 @miniufo
> Are you still planning on adding these instructions to the README and docs?

Oops, I forgot to add this part to the docs.  Now it is fixed at [here](https://xinvert.readthedocs.io/en/latest/Installation.html)
