### 2023-07-14 miniufo/xinvert#23 @miniufo
It is fixed now at [here](https://github.com/miniufo/xinvert/blob/master/docs/source/notebooks/05_reference_SWM.ipynb)
