### 2023-07-18 miniufo/xinvert#24 @NoraLoose
Re-running notebooks
When trying to re-run the notebooks, three of them fail:

- [x] In `11_Omega_equation.ipynb` the following fails because the netcdf file is not available
```
ds = xr.open_dataset('I:/Omega/OFES30_20011206_Qian/data.nc',
                     chunks={'lev':6}).astype('f4')
```

- [x] In `04_Eliassen_model.ipynb` the following fails:
```
ds = xr.open_dataset('../../../Data/Zonalmean.nc')
```
The latter is simply a typo in the filename; it needs to be changed to `./../../Data/ZonalMean.nc`

- [x] In `test_01_Stommel_Arons_model.ipynb` the following fails:
```
eps = xr.where(basin, eps, np.nan)
```
with the error
```
ValueError: cannot align objects with join='exact' where index/labels/sizes are not equal along these coordinates (dimensions): 'lon' ('lon',)
```

This is part of the review process at https://github.com/openjournals/joss-reviews/issues/5510.
