### 2023-07-18 22f03aa
Fix typos in software paper
