### 2023-07-18 bb1c1f6
Fix typos in software paper
