### 2023-07-19 miniufo/xinvert#24 @miniufo
Hi @NoraLoose thanks for your careful check.

>  In `11_Omega_equation.ipynb` the following fails because the netcdf file is not available

This notebook contains two examples.  The first one in the atmospheric context should be OK.  The second one use a large oceanic dataset (> 1GB) that cannot be uploaded to github (a 3D case).  It is not easy to reproduce the result by others because of this limit.  I put it here because I may sometimes check it myself, and I could also easily turn all the notebooks into docs.  Since data size cannot be reduced any further, I just don't know how to fix this.  Do you have any suggestion?

> In `04_Eliassen_model.ipynb` the following fails:

This can be fixed easily.

> In `test_01_Stommel_Arons_model.ipynb` the following fails:

All the notebooks start with `test` rather than a number are actually not ready for others to run.  So I guess I may move them to another place.  Only when they are ready, should I move them back here (I may need to modifiy the relative paths).
