### 2023-07-20 e6ac46f
fix issues in notebooks
