+++ docs/environment.yml
@@ -12,3 +12,4 @@ dependencies:
   - numba
   - proplot
   - cartopy
+  - pooch
