### 2023-07-20 miniufo/xinvert#24 @miniufo
Hi @NoraLoose, I've:
- fixed the large-file issue in omega example using `pooch`;
- fixed the filename in Eliassen example;
- move all the unfinished notebooks somewhere else.

Please check these again.  Thanks very much for the suggestions on solving these issues.
