### 2023-07-20 4513ad2
move environmental.yml to docs
