### 2023-07-25 921ddaa
Correct typos in software paper
