### 2023-07-26 #5510 @jbytecode
@miniufo - After applying the pull-request please consider changing/correcting the issues listed below

- In the last row of the table, in which a text of `**...** more problems **...**` included, should be removed.
- In the last stage, I see a part of

> `xinvert` is designed in a functional-programming (FP) style.  Users only need to import the function they are interested in and call it to get the inverted results (\autoref{table:1})

however, it is not true. Functional programming is a programming paradigm in which the functions are first-class citizen of the language, functions take functions as arguments, functions return functions, partial application, curriyng, immutability, etc. It is clear that being able to call functions by loading them from a package does not necessarily mean functional programming.

Maybe you want to say "xinvert is designed without using object oriented features of Python and only functions have been used", or is it better to say nothing about it?
