### 2023-07-26 #5510 @miniufo
Oh that is right.  What you mean here is more precise: I do not use object-oriented features but only define several functions which is not a functional style (only one function `animate_iteration` is more funtional-style).  I should correct these soon.
