### 2023-07-26 250b973
Update tests.yml
change typo
