+++ .github/workflows/tests.yml
@@ -21,7 +21,7 @@ jobs:
           miniconda-version: "latest"
           python-version: ${{ matrix.python-version }}
           environment-file: environment.yml
-          activate-environment: xgrads
+          activate-environment: xinvert
           auto-activate-base: false
 
       - name: Check workspace
