### 2023-07-26 #5510 @jbytecode
@miniufo - Could please put a comma in the bibtex file of

`doi       = {10.1357/002224020834442382}`

at the end of the line 117?
