+++ paper/paper.bib
@@ -114,7 +114,7 @@
   year      = 1954,
   publisher = {Ubiquity Press},
   pages     = {254-262},
-  doi       = {10.1357/002224020834442382}
+  doi       = {10.1357/002224020834442382},
   url       = {https://elischolar.library.yale.edu/journal_of_marine_research/818/},
 }
 
