+++ .github/workflows/tests.yml
@@ -13,7 +13,7 @@ jobs:
 
     steps:
       - name: Checkout source
-        uses: actions/checkout@v2
+        uses: actions/checkout@v3
 
       - name: setup conda environment
         uses: conda-incubator/setup-miniconda@v2
