### 2023-07-28 0ce80e2
remove 3.7 in workflow tests
