### 2023-07-28 #5510 @miniufo
> In the last row of the table, in which a text of **...** more problems **...** included, should be removed.

This is corrected.  Please see [here](https://github.com/miniufo/xinvert/blob/master/paper/paper.md#summary)

> however, it is not true. Functional programming is a programming paradigm in which the functions are first-class citizen of the language, functions take functions as arguments, functions return functions, partial application, curriyng, immutability, etc. It is clear that being able to call functions by loading them from a package does not necessarily mean functional programming.

This is corrected.  Please see [here](https://github.com/miniufo/xinvert/blob/master/paper/paper.md#usage)

> Could you please put a comma in the bibtex file at the end of line 117:

This is corrected.  Please see [here](https://github.com/miniufo/xinvert/blob/0ce80e254ec192a6e56b37596680f8fe0303c7f2/paper/paper.bib#L117C44-L117C44)

Please tell me if I missed some points @jbytecode .  Thank you very much for your help.
