### 2023-07-28 #5510 @jbytecode
@miniufo - The paper and the bibtex look fine. Could you please generate a new version tag and register the latest form of the repository as described in https://github.com/openjournals/joss-reviews/issues/5510#issuecomment-1650502230
