### 2023-07-28 #5510 @miniufo
@jbytecode Yes, I can generate a new tag but there is a problem.  I have sent a proposal to get a funding for this python package.  But it may take time to see if I finally get the funding.  Just Wondering if I still have chance to add the funding number in the future (e.g., like a proof reading in the later time so I can modify the funding part in the paper)?
