### 2023-07-28 #5510 @jbytecode
@miniufo - How much time do you need? You can add the funding number before I recommend an acceptance. You can now go on with the versioning, tagging and registering things independently of the paper. After everything seems okay (including addition of the funding number), please ping me again.
