+++ xinvert/__init__.py
@@ -31,4 +31,4 @@ from .utils import loop_noncore
 
 from .finitediffs import FiniteDiff, deriv, deriv2, padBCs
 
-__version__ = "0.1.2"
+__version__ = "0.1.3"
