### 2023-09-05 ec65264
update for JOSS paper
