+++ .readthedocs.yaml
@@ -1,0 +1,21 @@
+version: 2
+
+
+build:
+  os: ubuntu-22.04
+  tools:
+    python: "3.11"
+
+
+sphinx:
+  configuration: docs/conf.py
+  # You can configure Sphinx to use a different builder, for instance use the dirhtml builder for simpler URLs
+  # builder: "dirhtml"
+  # Fail on all warnings to avoid broken references
+  # fail_on_warning: true
+
+
+python:
+  install:
+    - requirements: docs/requirements.txt
+    - method: setuptools
