+++ .readthedocs.yaml
@@ -19,3 +19,5 @@ python:
   install:
     - requirements: docs/requirements.txt
     - method: setuptools
+      path: .
+
