+++ .readthedocs.yaml
@@ -8,7 +8,7 @@ build:
 
 
 sphinx:
-  configuration: docs/conf.py
+  configuration: docs/source/conf.py
   # You can configure Sphinx to use a different builder, for instance use the dirhtml builder for simpler URLs
   # builder: "dirhtml"
   # Fail on all warnings to avoid broken references
