### 2023-09-05 #5510 @miniufo
Hi @jbytecode, I've changed the funding number.  Since the conda has merged my PR, I also added a conda-install to the docs.  Now I've released v0.1.5 of the package.  Please check if everything is OK.
