### 2023-09-06 #5510 @jbytecode
@miniufo - Also the license should match the original value defined in the github repository
