### 2023-09-06 #5510 @jbytecode
@miniufo - Now the submission and the repository look good to me. I am recommending an acceptance. Our track editor will make the final decision. Thank you for sending your work to JOSS.

@DamienIrving, @NoraLoose - Thank you for reviewing this submission. Without you, we couldn't have done this job on our own.
