### 2023-09-06 #5510 @miniufo
Thank you very much @jbytecode @DamienIrving @NoraLoose for your comments and suggestions that make xinvert much much better.  The final PDF looks good to me.
